add a new mode days-interval
In this new mode of the date picker, the user can first select a date
and then select another date to form a date interval.
All the days in between are highlighted to provide visual feedback to
the user.

Note: the ISO format for dates interval is dates separated by '/' (e.g.
2012-01-01/2012-03-02)